feat(tests): improve test structure per review-tests.md findings
Sprint 145 (PBI-172): Relocate smoke tests
- Move 11 smoke tests from tests/test_runtime_coordinator_api.rs
- Tests now live alongside coordinator.rs in #[cfg(test)] module
- Reduces test pyramid imbalance (~9% fewer integration tests)

Sprint 146 (PBI-173): Parameterize tests with rstest
- Consolidate 3 duplicate offset clamping tests into 1 parameterized test
- Add rstest 0.23 dependency for parameterization support
- Named test cases: past_eof, start_past_end, row_past_eof

Sprint 147 (PBI-174): API visibility audit
- Make try_load_language_by_id private (only used internally)
- All other pub methods have external callers, visibility justified
