+++ Cargo.toml
@@ -46,6 +46,7 @@ criterion = { version = "0.5", features = ["html_reports"] }
 filetime = "0.2"
 insta = { version = "1", features = ["json"] }
 serial_test = "3"
+rstest = "0.23"
 
 [[bench]]
 name = "injection_tokens"
